drivers: display: ed2208: Fix format warning
%zu should only used for argument of size_t type, but the buf_size is of
uint32_t type. Casting to size_t is required. Otherwise, a warning will
occur with compiler of 64-bit platforms.

    drivers/display/display_ed2208_gca.c:238:25: error: format
    '%zu' expects argument of type 'size_t', but argument 3 has
    type 'uint32_t' {aka 'unsigned int'} [-Werror=format=]

+++ drivers/display/display_ed2208_gca.c
@@ -235,7 +235,7 @@ static int ed2208_gca_write(const struct device *dev, const uint16_t x, const ui
 	}
 
 	if (buf == NULL || desc->buf_size < buf_len) {
-		LOG_ERR("Invalid buffer: %p (%zu < %zu)", buf, desc->buf_size, buf_len);
+		LOG_ERR("Invalid buffer: %p (%zu < %zu)", buf, (size_t)desc->buf_size, buf_len);
 		return -EINVAL;
 	}
